✨feat(Text Components): Finishing large expressions.
