+++ src/components/TextRender/NumberedListComponent/NumberedListComponent.tsx
@@ -1,15 +1,21 @@
 import { NumericText, StyledNumberedListComponent } from "./styles";
 
 interface INumberedListComponent {
-  list: string[];
+  list?: string[];
 }
 
 function NumberedListComponent({ list }: INumberedListComponent) {
+  console.log(list);
   return (
     <StyledNumberedListComponent className="numeric-list-component">
-      {list.map((item, index) => {
+      {list?.map((item, index) => {
         return (
-          <NumericText data-numeric-index={index} className="numeric-list-item"  title={item} key={index}>
+          <NumericText
+            data-numeric-index={index}
+            className="numeric-list-item"
+            title={item}
+            key={index}
+          >
             {item}
           </NumericText>
         );
