+++ src/shared/utils/TextTransform.ts
@@ -11,15 +11,8 @@ import { HighlightBoldSubTitleComponent } from "../../components/TextRender/High
 import { BlockQuoteComponent } from "../../components/TextRender/BlockQuoteComponent/BlockQuoteComponent";
 import { CodeBlockComponent } from "../../components/TextRender/CodeBlockComponent/CodeBlockComponent";
 import { BulletListComponent } from "../../components/TextRender/BulletListComponent/BulletListComponent";
-
-type UnionExpressions = LargeExpressions | StartExpressions;
-
-type ExpressionType = {
-  text: string | string[];
-  type: UnionExpressions;
-  start?: number;
-  end?: number;
-};
+import { ExpressionType, mapExpressions } from "./Expressions";
+import { NumberedListComponent } from "../../components/TextRender/NumberedListComponent/NumberedListComponent";
 
 /**
  * Splits the input text into an array of lines.
@@ -44,7 +37,7 @@ function getLines(text: string): string[] {
  * @param {string[]} textInArray - An array of strings to be processed for start expressions.
  * @returns {React.Component[]} An array of React components mapped from the start expressions in the input array.
  */
-function getExpressions(textInArray: string[]) {
+function mountExpressions(textInArray: string[]) {
   const expressions: ExpressionType[] = mapExpressions(textInArray);
 
   // Map each object containing the identified starting expression and text to a React component.
@@ -52,190 +45,11 @@ function getExpressions(textInArray: string[]) {
     expressionsMapByComponent(expression, index)
   );
 
-  console.log({ expressions, components });
-
   // Return the array of React components.
-  return components;
-}
-
-function mapExpressions(textInArray: string[]) {
-  const expressions: ExpressionType[] = [];
-
-  let currentIndex = 0;
-
-  while (currentIndex < textInArray?.length) {
-    try {
-      const line = textInArray[currentIndex];
-
-      const findStartExpressions = searchStartExpression(line);
-      const findLargeExpressions = searchLargeExpression(line);
-
-      const isStartExpression = findStartExpressions?.length > 0;
-      const isLargeExpressions = findLargeExpressions?.length > 0;
-
-      if (!isStartExpression && !isLargeExpressions) {
-        expressions.push({
-          text: line,
-          type: StartExpressions.Text,
-        });
-      }
-
-      if (isStartExpression) {
-        const type = findStartExpressions[0];
-
-        const startExpression = convertForStartExpressions(type, line);
-        expressions.push(startExpression);
-
-        continue;
-      }
-
-      if (isLargeExpressions) {
-        const type = findLargeExpressions[0];
-
-        const largeExpression = convertLargeExpressions(
-          type,
-          line,
-          textInArray,
-          currentIndex
-        );
-
-        if (
-          typeof largeExpression?.start === "number" &&
-          typeof largeExpression?.end === "number"
-        ) {
-          textInArray = [
-            ...textInArray.slice(0, largeExpression.start),
-            ...textInArray.slice(largeExpression.end),
-          ];
-        }
-
-        expressions.push(largeExpression);
-
-        continue;
-      }
-    } catch (error) {
-      console.error(error);
-    } finally {
-      currentIndex++;
-    }
-  }
-
-  return expressions;
-}
-
-function searchStartExpression(text: string): StartExpressions[] {
-  // Define an array of possible start expressions.
-  const startExpressions = [
-    StartExpressions.Title,
-    StartExpressions.Subtitle,
-    StartExpressions.BlockQuote,
-    StartExpressions.LargeBoldSubTitle,
-    StartExpressions.MediumBoldSubTitle,
-    StartExpressions.SmallBoldSubTitle,
-    StartExpressions.HighlightBoldSubTitle,
-    StartExpressions.Text,
-    StartExpressions.BreakLine,
-  ];
-
-  // Filter the startExpressions array to find the expression(s) that match the start of the input text.
-  const matchedExpressions = startExpressions.filter((expression) => {
-    const textBySpaces = text.split(" ");
-
-    // Check if the first word in the text matches the current expression.
-    return textBySpaces[0] === expression;
-  });
-
-  return matchedExpressions;
-}
-
-function searchLargeExpression(text: string): LargeExpressions[] {
-  const largeExpression = [
-    LargeExpressions.BulletList,
-    LargeExpressions.NumberedList,
-    LargeExpressions.CodeBlock,
-  ];
-
-  // Filter the startExpressions array to find the expression(s) that match the start of the input text.
-  const matchedExpressions = largeExpression.filter((expression) => {
-    const textBySpaces = text.split(" ");
-
-    // Check if the first word in the text matches the current expression.
-    return textBySpaces[0] === expression;
-  });
-
-  return matchedExpressions;
-}
-
-/**
- * Checks the start of a text to identify the type of starting expression.
- *
- * @param {string} text - The input text to be checked for starting expressions.
- * @returns {{ text: string; expression: StartExpressions; }} An object containing the identified starting expression type and the remaining text after removing the expression.
- */
-function convertForStartExpressions(
-  type: StartExpressions,
-  text: string
-): ExpressionType {
-  // Remove the identified expression from the start of the text and trim any leading spaces.
-  const textWithoutExpression = removeExpressionFromText(text, type);
-
-  // Return an object with the identified starting expression and the remaining text after removing the expression.
   return {
-    text: textWithoutExpression,
-    type,
+    expressions,
+    components,
   };
-}
-
-function convertLargeExpressions(
-  type: LargeExpressions,
-  text: string,
-  textInArray: string[],
-  currentIndex: number
-): ExpressionType {
-  const index = currentIndex + 1;
-  const startArray = textInArray.slice(index);
-
-  if (type === LargeExpressions.CodeBlock) {
-    const findCloseCodeBlock = startArray.indexOf(LargeExpressions.CodeBlock);
-
-    if (findCloseCodeBlock === -1) {
-      return {
-        text,
-        type: StartExpressions.Text,
-      };
-    }
-
-    return {
-      text: textInArray.slice(index, index + findCloseCodeBlock).join("\n"),
-      type,
-      start: currentIndex,
-      end: index + findCloseCodeBlock,
-    };
-  }
-
-  if (type === LargeExpressions.BulletList) {
-    const findNextBreakLine = startArray.indexOf(StartExpressions.BreakLine);
-
-    if (findNextBreakLine === -1) {
-      return {
-        text,
-        type: StartExpressions.Text,
-      };
-    }
-
-    const list = textInArray
-      ?.slice(currentIndex, currentIndex + findNextBreakLine + 1)
-      ?.map((text) => removeExpressionFromText(text, type));
-
-    return {
-      text: list,
-      type,
-      start: currentIndex,
-      end: index + findNextBreakLine + 1,
-    };
-  }
-
-  return {} as ExpressionType;
 }
 
 /**
@@ -258,7 +72,6 @@ function expressionsMapByComponent(expression: ExpressionType, index: number) {
   // Use a switch statement to determine the type of starting expression and map it to the corresponding React component.
   switch (type) {
     case StartExpressions.Title:
-      // If the expression is Title, create a TitleComponent with the provided text and default props.
       return createComponent(TitleComponent, defaultProps);
 
     case StartExpressions.Subtitle:
@@ -300,14 +113,15 @@ function expressionsMapByComponent(expression: ExpressionType, index: number) {
         list: text,
       });
 
+    case LargeExpressions.NumberedList:
+      return createComponent(NumberedListComponent, {
+       ...defaultProps,
+        list: text,
+      });
+
     default:
       return createComponent(TextComponent, defaultProps);
   }
 }
 
-function removeExpressionFromText(text: string, type: UnionExpressions) {
-  const textWithoutExpression = text.replace(type, "").trimStart();
-  return textWithoutExpression;
-}
-
-export { getLines, getExpressions };
+export { getLines, mountExpressions };
